:pencil2: Add .gitignore file
This commit includes the addition of a .gitignore file to the repository.
This .gitignore file contains common ignores for different environments
and languages, such as Linux, macOS, Python, Svelte, and Windows.
This should help in keeping the repository clean from unnecessary files
and folders that are commonly generated during development and should not
be included in the repository.