:sparkles: Add frontend structure and files
This commit introduces a significant amount of new files necessary for the frontend
development of our project. These include configuration files for eslint, npm,
gitignore, as well as source files for the frontend in the /src directory, and
static files.

The reason for this massive addition is to set up the initial structure of our
frontend, which uses Svelte. The frontend now includes several pages and
components, as well as their typescript definitions and routing configurations.

For styles, we added styles.css which will hold all our global styles, and the
static directory now holds our static files like favicon and robots.txt.

Furthermore, this commit includes a sverdle game, with server functions, routes
and a guide on how to play. The game is still a work in progress and some routes
might not work as intended yet. But this will give a solid foundation for
subsequent development.