:shower: Refactor frontend and remove obsolete files
The frontend has been refactored to improve the overall user experience.
Several files, such as `Counter.svelte`, `Header.svelte`, and `+page.svelte`,
were obsolete and hence removed.

This commit also includes modification of the layout and styling of the frontend
as reflected in `+layout.svelte` and `styles.css` files. The modification will
give the website a more modern look and will make it easier for users to navigate.

It is believed that this change will contribute to the maintainability and scalability
of our application in the future.