:sparkles: Implement backend Dockerization and fix frontend settings
This commit introduces several changes:

First, Dockerization of the backend is accomplished with a new Dockerfile, allowing the backend to run in an isolated environment. This addition includes setting up a Python environment and copying over necessary files, improving the portability and reproducibility of the backend server.

Second, the addition of a new FastAPI server setup file 'main.py' and its 'requirements.txt' file has been made, establishing the backend server and its necessary packages. The FastAPI server will be used for handling various API requests.

Third, modifications to 'compose.yml', 'default.conf', and 'package.json' are made. These changes involve the command execution syntax for npm, proxy header settings, and the host configuration for Vite in the frontend. These changes are made to improve the interaction between the frontend and backend and to fix minor bugs in the setup.

The main goal of these changes is to set up a working Dockerized backend and make some minor fixes to the frontend setup. This will improve the development environment and make it easier for other developers to contribute to the project.