:bug::+1: Fix websocket issues and improve networking
This commit does two major things. First, it addresses some bugs with the websocket handling in 'backend/main.py'. The connected_clients data structure was previously a set, which is not suitable for this use case. The code was changed to use a list instead, which is more appropriate. Additionally, the data reception method was changed from 'receive_json()' to 'receive_text()'.

Second, this commit improves networking in both the backend 'default.conf' file and the frontend '+page.svelte' file. In the 'default.conf' file, the formatting was fixed to be more consistent. In the '+page.svelte' file, the WebSocket endpoint was changed from 'localhost' to '0.0.0.0', expanding the range of accessible hosts.

These changes help to improve the stability and usability of the application.