:sparkles: Refactor WebSocket functionality and Improve UI
This commit primarily does two things:

Firstly, the WebSocket functionality is refactored in order to make the code more efficient and readable. The 'websocket.ts' file is split into 'openSocket.ts' and 'sendImage.ts', and unnecessary conditions in the ImageUploader.svelte are simplified. The way the WebSocket connection is opened in '+page.svelte' is also changed to only send images after the connection is open.

Secondly, the UI in 'ImageList.svelte' is improved by displaying the name of the image before the image itself. This will allow users to know the name of the image they are viewing.

The purpose of these changes is to enhance the user experience by making the UI more intuitive and the WebSocket functionality more efficient. The split of 'websocket.ts' into 'openSocket.ts' and 'sendImage.ts' also helps in separating the concerns and improving the readability of the code.