:sparkles: :bug: Improve image upload feature and fix bugs
This commit addresses two main areas. Firstly, the image upload feature in the backend (main.py) has been enhanced by refactoring the upload_endpoint function to support base64 encoding and decoding, enabling better handling of image data. As a part of this feature, an additional package "Pillow" has been added to the project requirements.

The second aspect tackled in this commit is the bug fixing in the frontend part of the application, in which the 'sendImage.ts' file has been renamed to 'sendChunk.ts', and the function 'sendImage' has been renamed to 'sendChunk' to better reflect its purpose.

The logic in the '+page.svelte' file has also been updated to work with the new 'sendChunk' function, providing an improved user experience.

This refactoring and bug fixing are intended to provide a more robust image upload feature and improve the overall user experience. It also sets the stage for further possible enhancements such as providing user feedback during image upload.