:pencil2: Add MIT License
This commit introduces the MIT License to the project. This license grants permissions to any person obtaining a copy of this software and the associated documentation files, allowing them to freely use, copy, modify, merge, publish, distribute, sublicense, and sell copies of the software. All recipients of the software must include this license in all copies or substantial portions of the software.

Adding an MIT License is beneficial for fostering an open-source community, as it provides clear permissions and conditions to users and contributors. It ensures that contributors retain copyright rights while granting others the mentioned permissions.